Move parsing to dedicated file
